+++ R/create_env.R
@@ -4,7 +4,7 @@
 #'   by `run()`.
 #'
 #' @param packages Character vector. Names of the packages, and
-#'   version strings if necessary, e.g. 'python=3.11'. The use of the `packages`
+#'   version strings if necessary, e.g. 'python=3.13'. The use of the `packages`
 #'   argument assumes that env_file is not used.
 #'
 #' @param env_file Character. Path to the YAML file with Conda Environment